+++ geo-types/src/line_string.rs
@@ -102,6 +102,7 @@ where
     T: CoordinateType;
 
 /// A `Point` iterator returned by the `points_iter` method
+#[derive(Debug)]
 pub struct PointsIter<'a, T: CoordinateType + 'a>(::std::slice::Iter<'a, Coordinate<T>>);
 
 impl<'a, T: CoordinateType> Iterator for PointsIter<'a, T> {
@@ -184,7 +185,7 @@ impl<T: CoordinateType> LineString<T> {
     pub fn close(&mut self) {
         if !self.is_closed() {
             // by definition, we treat empty LineString's as closed.
-            debug_assert!(self.0.len() > 0);
+            debug_assert!(!self.0.is_empty());
             self.0.push(self.0[0]);
         }
     }